549 move ruler from number.py
